requirements.txt
streamlit
mysql-connector-python
apscheduler
streamlit-autorefresh
dnspython
